+++ wcomponents-core/src/main/java/com/github/bordertech/wcomponents/WApplication.java
@@ -2,6 +2,11 @@ package com.github.bordertech.wcomponents;
 
 import com.github.bordertech.wcomponents.util.Config;
 import com.github.bordertech.wcomponents.util.I18nUtilities;
+import com.github.bordertech.wcomponents.util.Util;
+import java.io.Serializable;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * <p>
@@ -17,7 +22,7 @@ import com.github.bordertech.wcomponents.util.I18nUtilities;
 public class WApplication extends AbstractMutableContainer implements AjaxTarget, DropZone {
 
 	/**
-	 * The application icon url parameter.
+	 * The application icon url.
 	 */
 	private static final String ICON_URL = Config.getInstance().getString(
 			"bordertech.wcomponents.application.icon.url");
@@ -100,6 +105,7 @@ public class WApplication extends AbstractMutableContainer implements AjaxTarget
 	@Override
 	public String getNamingContextId() {
 		boolean append = isAppendID();
+
 		if (!append) {
 			// Check if this is the top level name context
 			NamingContextable top = WebUtilities.getParentNamingContext(this);
@@ -108,6 +114,7 @@ public class WApplication extends AbstractMutableContainer implements AjaxTarget
 				append = true;
 			}
 		}
+
 		if (append) {
 			return getId();
 		} else {
@@ -137,6 +144,154 @@ public class WApplication extends AbstractMutableContainer implements AjaxTarget
 	public static WApplication instance(final WComponent base) {
 		WApplication appl = WebUtilities.getClosestOfClass(WApplication.class, base);
 		return appl;
+	}
+
+	/**
+	 * Add custom javaScript located at the specified URL to be used by the Application.
+	 *
+	 * @param url URL to a javaScript resource
+	 * @return the application resource the URL details are held in
+	 */
+	public ApplicationResource addJsUrl(final String url) {
+		if (Util.empty(url)) {
+			throw new IllegalArgumentException("A URL must be provided.");
+		}
+		ApplicationResource res = new ApplicationResource(url);
+		addJsResource(res);
+		return res;
+	}
+
+	/**
+	 * Add custom javaScript held as an internal resource to be used by the application.
+	 *
+	 * @param fileName the javaScript file name.
+	 * @return the application resource the resource details are held in
+	 */
+	public ApplicationResource addJsFile(final String fileName) {
+		if (Util.empty(fileName)) {
+			throw new IllegalArgumentException("A file name must be provided.");
+		}
+		InternalResource resource = new InternalResource(fileName, fileName);
+		ApplicationResource res = new ApplicationResource(resource);
+		addJsResource(res);
+		return res;
+	}
+
+	/**
+	 * Add a custom javaScript resource that is to be loaded and used by the application.
+	 *
+	 * @param resource the javaScript resource
+	 */
+	public void addJsResource(final ApplicationResource resource) {
+		WApplicationModel model = getOrCreateComponentModel();
+		if (model.jsResources == null) {
+			model.jsResources = new HashSet<>();
+		}
+		model.jsResources.add(resource);
+	}
+
+	/**
+	 * Remove all custom javaScript resources.
+	 */
+	public void removeAllJsResources() {
+		getOrCreateComponentModel().jsResources = null;
+	}
+
+	/**
+	 * Remove a custom javaScript resource.
+	 *
+	 * @param resource the javaScript resource to remove.
+	 */
+	public void removeJsResource(final ApplicationResource resource) {
+		WApplicationModel model = getOrCreateComponentModel();
+		if (model.jsResources != null) {
+			model.jsResources.remove(resource);
+		}
+	}
+
+	/**
+	 * @return the set of custom javaScript application resources
+	 */
+	public Set<ApplicationResource> getJsResources() {
+		WApplicationModel model = getComponentModel();
+		if (model.jsResources == null) {
+			return Collections.EMPTY_SET;
+		}
+		return Collections.unmodifiableSet(model.jsResources);
+	}
+
+	/**
+	 * Add custom CSS located at the specified URL to be used by the Application.
+	 *
+	 * @param url URL to a CSS resource
+	 * @return the application resource the URL details are held in
+	 */
+	public ApplicationResource addCssUrl(final String url) {
+		if (Util.empty(url)) {
+			throw new IllegalArgumentException("A URL must be provided.");
+		}
+		ApplicationResource res = new ApplicationResource(url);
+		addCssResource(res);
+		return res;
+	}
+
+	/**
+	 * Add custom CSS held as an internal resource to be used by the Application.
+	 *
+	 * @param fileName the CSS file name.
+	 * @return the application resource the resource details are held in
+	 */
+	public ApplicationResource addCssFile(final String fileName) {
+		if (Util.empty(fileName)) {
+			throw new IllegalArgumentException("A file name must be provided.");
+		}
+		InternalResource resource = new InternalResource(fileName, fileName);
+		ApplicationResource res = new ApplicationResource(resource);
+		addCssResource(res);
+		return res;
+	}
+
+	/**
+	 * Add a custom CSS resource that will be loaded and used by this application.
+	 *
+	 * @param resource the CSS resource
+	 */
+	public void addCssResource(final ApplicationResource resource) {
+		WApplicationModel model = getOrCreateComponentModel();
+		if (model.cssResources == null) {
+			model.cssResources = new HashSet<>();
+		}
+		model.cssResources.add(resource);
+	}
+
+	/**
+	 * Remove all custom CSS resources.
+	 */
+	public void removeAllCssResources() {
+		getOrCreateComponentModel().cssResources = null;
+	}
+
+	/**
+	 * Remove a custom CSS resource.
+	 *
+	 * @param resource the CSS resource to remove.
+	 */
+	public void removeCssResource(final ApplicationResource resource) {
+		WApplicationModel model = getOrCreateComponentModel();
+		if (model.cssResources != null) {
+			model.cssResources.remove(resource);
+		}
+	}
+
+	/**
+	 * @return the list of custom CSS resources
+	 */
+	public Set<ApplicationResource> getCssResources() {
+		WApplicationModel model = getComponentModel();
+		if (model.cssResources == null) {
+			return Collections.EMPTY_SET;
+		}
+		return Collections.unmodifiableSet(model.cssResources);
 	}
 
 	/**
@@ -203,5 +358,107 @@ public class WApplication extends AbstractMutableContainer implements AjaxTarget
 		 * Flag to append the Application ID to the children IDs.
 		 */
 		private boolean appendID;
+
+		/**
+		 * Set of custom javaScript resources.
+		 */
+		private Set<ApplicationResource> jsResources;
+
+		/**
+		 * Set of custom CSS resources.
+		 */
+		private Set<ApplicationResource> cssResources;
 	}
+
+	/**
+	 * Holds the details of CSS and JS custom resources.
+	 */
+	public static class ApplicationResource implements Serializable {
+
+		/**
+		 * Unique resource id.
+		 */
+		private final String resourceId;
+
+		/**
+		 * Resource URL.
+		 */
+		private final String url;
+
+		/**
+		 * Internal resource.
+		 */
+		private final InternalResource resource;
+
+		/**
+		 * @param url URL to a resource
+		 */
+		public ApplicationResource(final String url) {
+			if (Util.empty(url)) {
+				throw new IllegalArgumentException("A URL must be provided.");
+			}
+			this.resourceId = "url:" + url;
+			this.url = url;
+			this.resource = null;
+		}
+
+		/**
+		 * @param resource internal resource
+		 */
+		public ApplicationResource(final InternalResource resource) {
+			if (resource == null) {
+				throw new IllegalArgumentException("A resource must be provided.");
+			}
+			this.resourceId = "file:" + resource.getResourceName();
+			this.url = null;
+			this.resource = resource;
+		}
+
+		/**
+		 * @return the resource id
+		 */
+		public String getResourceId() {
+			return resourceId;
+		}
+
+		/**
+		 * @return the resource URL
+		 */
+		public String getUrl() {
+			return url;
+		}
+
+		/**
+		 *
+		 * @return the internal resource
+		 */
+		public InternalResource getResource() {
+			return resource;
+		}
+
+		/**
+		 * @return the target URL of the resource
+		 */
+		public String getTargetUrl() {
+			return url == null ? resource.getTargetUrl() : url;
+		}
+
+		/**
+		 * {@inheritDoc}
+		 */
+		@Override
+		public boolean equals(final Object obj) {
+			return obj instanceof ApplicationResource && resourceId.equals(((ApplicationResource) obj).getResourceId());
+		}
+
+		/**
+		 * {@inheritDoc}
+		 */
+		@Override
+		public int hashCode() {
+			return resourceId.hashCode();
+		}
+
+	}
+
 }